removed test_inlinepanel_interity.py and move tests in test_edit_page.py
